refactor the code and splid the js code
