+++ src/variables.js
@@ -1,0 +1,8 @@
+
+export const baseUrl =
+  'https://us-central1-js-capstone-backend.cloudfunctions.net/api';
+
+export const game = {
+  name: 'tower books',
+  gameId: 'CLwRs3Mgq5HE7kyDdKH0',
+};
